Add `Autocomplete` component.
`TripDay` interface now has 'locations' array, which is just a string array instead of an `activities` array. This is much simpler for testing purposes.

Add base code for location suggestions. Currently it isn't fully working as I am temporarily banned from the service.

Add map on the right hand side of the screen.