Major feature update:
The user can now create a new trip which shows up on the home page.

There is now a router throughout the app. I am using the 'Wouter' package.

There is a default list of trips which come from `TripDB`. This is a simple in-memory database and is not persistent.

[...]